Splitting tonemapper into its own class
