+++ servers/rendering/renderer_rd/shaders/effects/SCsub
@@ -15,5 +15,3 @@ if "RD_GLSL" in env["BUILDERS"]:
     # compile shaders
     for glsl_file in glsl_files:
         env.RD_GLSL(glsl_file)
-
-SConscript("effects/SCsub")
