+++ servers/rendering/renderer_rd/shaders/effects/tonemap.glsl
